Refactor: extract multi-action button into a reusable component.
